fix deep recursion on self referential schemas
either directly or indirectly[1] by maintaining a depth count and
throwing warnings and exiting early should be blow the stack. this
can be changed as necessary using the max_depth constructor arg

add a test to check for deep recursion. update perldoc and bump
VERSION and Changes for CPAN release

[1] see https://api.opentrials.net/v1/swagger.yaml for an example